refactor: bring in modules, create file with API constants, move class for API resquests to separate file
