+++ marvel-comics/src/constants/api.js
@@ -1,0 +1,6 @@
+// Константы для работы с API
+export const API_KEY = 'a5837db97d72016c81a7a776f4240db9';
+// Т.к у всех URL общий базовый путь, можно вынести его в отдельную константу, с которым будем конкатенировать новые пути
+export const API_URL = 'https://gateway.marvel.com/v1/public/';
+export const URL_COMICS = 'comics';
+export const URL_CHARACTERS = 'characters';
